Add files via upload
This area chart shows the number of new cases by day. In the title, I added the total number of new cases, and below that is the present percent difference between days. This calculation varies from day to day while I just calculated for the latest vs the previous day.